requirements-testing.txt
factory-boy==2.6.0
pytest==2.9.1
pytest-cov==2.3.0
python-coveralls==2.7.0
pytest-django==2.9.1
responses==0.5.1